+++ metadata.rb
@@ -6,7 +6,7 @@ maintainer_email 'j@p4nt5.com'
 license          'apache2'
 description      'Installs/Configures Divvy'
 long_description 'Installs/Configures Divvy'
-version          '0.2.1'
+version          '0.2.2'
 
 depends          'mac-app-store', '~> 1.0'
 depends          'privacy_services_manager', '~> 1.0'