+++ metadata.rb
@@ -6,7 +6,7 @@ maintainer_email 'j@p4nt5.com'
 license 'apache2'
 description 'Installs/Configures Divvy'
 long_description 'Installs/Configures Divvy'
-version '0.2.2'
+version '1.0.0'
 
 source_url 'https://github.com/roboticcheese/divvy-chef'
 issues_url 'https://github.com/roboticcheese/divvy-chef/issues'