+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Divvy Cookbook CHANGELOG
 ========================
 
+Unreleased
+----------
+
 v0.2.1 (2016-02-22)
 -------------------
 - Fix issue failing to enable service on Windows